First working version of Index and Extract!
Made indexing and extraction work. However, to make things work, we need
large indices. This is because we need to store a dictionary for each
entry in the index. This is currently of size 32kByte and can be reduced
by using zlibs compress2/uncompress methods by around 60%.

This is not a final version and needs a lot more heavy grinding, e.g:
- Code cleanups and unification
- More tests and checks on startup
- Improve the CLI interface
  - Change access to FASTQ record access (currently it is working on
    lines, which might still be kept, but the main purpose of this tool
    is FASTQ processing)
- Improve code to use (much) less memory
(Just need to work on different things atm, so at least push it to the
 repo)

Raise minimum cmake to 3.13

Also tried to upgrade Boost from 1.54 to 1.69, which horribly failed.
Methods like e.g. exists(path), is_symlink(path) left horrible pointer
mess in calling objects. Though we could track this down a bit, we do
not know, where the real bug source is located.

/////// Source code

In CommonStructsAndConstants:
- Move several constants to CommonStructsAndConstants.
- Add equality (== !=) not (!) and bool() operators to IndexHeader
- In IndexEntryV1:
  * Layout changes from relative to absolute values.
  * Add the blockID() field to IndexEntryV1.
  * Rename relativeBlockOffsetInRawFile to blockOffsetInRawFile.
  * Change types for offsets from ushort to ulong.
  * Add the 32kByte sized dictionary field which is needed for inflate
    init in extraction mode. This value should be compressed in a future
    release (ca 60% storage decrease). The field
    compressedDictionarySize is already planned but not functional.

Added the IndexLine struct, which holds a decompressed and processed
IndexEntry. Also this is currently brute force and the whole index is
loaded fully to memory. This will need to change in the future to
decrease memory consumption.

The Starter will now create an ExtractorRunner with proper startline and
linecount.

In ActualRunner, extend checkPremises() with more checks. Also here, we
ran into a lot of trouble when trying to upgrade Boost from 1.54 to 1.69

In IndexerRunner, change indexer to Indexer* type and delete in in the
constructor. This and the other changes in the class were done due to
the upgrade conflicts of Boost.

In ExtractorRunner:
- Add checkPremises() and run()
- checkPremises() will check the Runner premises + the used extractor
  premises.
- run() will call extractor->extractReadsToCout()

In IndexReader:
- Add comments.
- Add the readerIsOpen field.
- Add the readHeader field and make readIndexHeader() private.
- Remove the create() method and make the class instantiable via a
  public constructor.
- Rename open() to tryOpenAndReadHeader() and make it public. Add checks
  and error messages.
- Add readIndexFile() method. This will most likely be removed in future
  versions with a more intelligent index loader approach.
- Add readIndexFileV1() method an comments for V2/V3

In Indexer:
- This class really needs to be optimized!
- Remove constants and use global constants.
- Add the calculateIndexBlockInterval() which will calculate some
  "smart" interval depending on the FASTQ file size.
- Change a lot of code. Tried to put it comments, but this really needs
  improvement.

In Extractor:
- First: Added the class
- Copied some code from Indexer, this should later be merged somehow.
- Most important: Add the extractReadsToCout() method which currently
  lacks checks. The method will jump into the gz file, initialize the
  inflater, search for the first line to extract and then extract until
  EOF or extractedLines == lineCount. Also here, the extractor is based
  on lines rather than records. This should be changed.

/////// Tests

Change a lot of the tests to match new constructors.

In CommonStructsTest:
- Change tests for IndexHeader and IndexEntry to match new sizes.
- Add tests for new operators()

Extend ExtractorRunnerTest to cover one more test case.

In IndexerTest:
- Add test for CalculateBlockInterval.
- Change most tests so, that the Indexer is created with -1 for the
  block interval to request block interval calculation.
  (Test create large index will use a block interval of 4.)
- Extend test create large index to check some index entries for
  validity. Values for this were calculated using the original zran.c
  code. Also put in an extraction test which compares decompressed with
  original data.

Add ExtractorTest:
- Add two non test methods runRangedExtractionTest() and
  runComplexDecompressionTest() which will be called by
  TEST_CREATE_EXTRACTOR_AND_EXTRACT_SMALL_TO_COUT to run a series of
  index / extract processes and compare extracted data to decompressed
  original data.

Extend StarterTest to use a complete set of parameters for extraction.