Various refactorings
Started to use "using namespace::class" in some cases.

Change a lot of minor things due to clang-tidy complaints.

Changed ulong to u_int64_t in the whole application to guarantee a
length of 8 Byte unsigned values. ulong itself is not guaranteed to be
of length 8 Byte (e.g. Microsoft says it is 4 Bytes long). Also changed
some of the uint's to u_int32_t though this is not necessary.

Changed the WINDOW_SIZE / CHUNK_SIZE to #defines instead of const uint,
with the old code, the chunk[CHUNK_SIZE]{0} syntax did not work and
needed an additional memset.

Renamed IndexLine to IndexEntry. IndexEntry is a decompressed and
processed index entry representation for all the versioned
IndexEntryV[n] structs. This is to docouple index reading and writing
from internal processes.

Added the toIndexEntry() method to IndexEntryV1.

Created the ZLibBasedFASTQProcessorBaseClass class as a mutually used
base class for the Indexer and Extractor classes. The class stores a
list of variables and methods used for both subclasses.

Cleanup Indexer and Extractor and make them more readable. Though a lot
more work could be done here. Also unify the buildIndex() and extract()
methods so they look a lot alike. This should be base for further
refactoring. Moved a lot of values from both classes to the new base
class.

The Extractor will now read in IndexEntry-wise until the proper
IndexEntry is found. This way, we do not need to read in the whole index
which makes sense for memory consumption. Big FASTQ files will have big
index files and the Extractor shall use only few memory.

Changed readIndexFileV1() to return a vector of IndexEntryV1 and create
the readIndexEntry() method.

/// Tests

Add ZLibBasedFASTQProcessorBaseClassTest file with a mock class to
enable testing of the class. Move the test for z_stream init to this
file.