Make some Starter tests work again (... with caveats)
Testing TCLAP (which is eventually done in the Starter tests) fails,
because TCLAP calls exit() effectively stopping all tests, when the
command line parameters cannot be parsed. This leads to the bad effect,
that we can only run valid tests but no tests with invalid command line
parameters.

To at least get a message, that something went wrong and that the test
binary exited unintentionally, I put in a switch after the tests are
finished. The switch is recognized by a trap to exit() and a message is
printed to the command line to indicate an error, if the switch is not
set properly.