Add example script for mass tests
The files are working on our LSF environments.

Basically you call fqimasstestLSF.sh with a list of FASTQ files and an
output directory to where the index files will be written.

The script fqitest.sh is called for each FASTQ. Note, that the script
will extract the FASTQ to the local scratch storage.

Keep in mind, that these scripts are more an example and will not
necessarily work on your platform.