Add timer methods, Fix several bugs
The Starter now parses the variables for extract BEFORE we access the
indexFile and debugMode parameters.

Extract a method to print an index entry to the console and put it in
IndexStatsRunner.

Output some minor statistics when extraction gets started.
Also fix a bug in the indexer. The first index entry was not stored,
when failsafe distance was on.

Add A LOT of build options for debug mode. These should be handled with
care to reduce warnings and increase code robustness.