Add S3 support
With this commit, S3 is basically working. However, we currently do not
know, how to properly end / abort a running download (e.g. needed for
extract). Also the system might seem a little complicated because we use
named pipes for async downloads / streaming. The reason for this is,
that we could not find a way to work directly with something like an
HTTP stream (or so).

The S3 addition needed a heavy rework of the I/O structure. This meant:
- Rename InputSource class to Source class
- Implementation of further Sink and Source classes
- Integration of Sink and Source everywhere in the code instead of the
  mixed Source/path combinations.
- New classes are IOBase, ConsoleSink, PathSink, S3Sink and S3Source

!!! Although S3 works, most of the tests for it don't. This is all still
incomplete and needs a lot more work. Also the Lock handling
(LockHandler, see below) is not yet solved.

Changed a lot of the #include directives to use a "full" path instead of
relative values, e.g.
  #include "../io/PathInputSource.h" => #i... "process/io/PathSource.h"
This is done all over the application.

Smaller updates to the README.

Move a lot of command line init code to the CMake files. Still need a
flag to allow optional build of S3 support.

In ErrorAccumulator add some more methods to output messages and one
method to merge three vectors.

Introduce the Sink class for for output classes (S3/Filesystem/Console)

IOHelper:
- Create methods createTempDir/File, createFIFO, checkFileReadability
  and fullPath

Introduce the LockHandler concept and base class. Currently, there is
only the PathLockHandler class available, which was formerly the
IndexProcessor class. There should be another LockHandler for S3 in the
future.

Add the S3Service and FQIS3Client classes, which are used to manage
setup and deinit of the S3 interface and to allow to run some generic S3
requests.

Refactored and unified the different CLIModeParser classes a lot. They
are now a lot more modular and readable.

Added some test classes and some basic tests

Changed the TestResourcesAndFunctions class to use the IOHelper features