Implement IndexEntry storage strategies for index creation
These strategies are used to decided whether an IndexEntry will be
stored or not. The current block based (safeguarded) approach will still
be usable and is encapsulated in the BlockDistanceStorageStrategy class.

Add the IndexEntryStorageStrategy base class.

Add the ByteDistanceStorageStrategy class to allow the user to set a
Byte value like 1T, 2g, 3m and 4K as a minimum distance between index
entries.

Rename the field offsetInRawFile to blockOffsetInRawFile in IndexEntry
to match IndexEntryV1.

Move the method calculateIndexBlockInterval from Indexer to the new
IndexEntryStorageStrategy class.

Add a test class for the new strategy classes.