Reworked S3 GetObject by moving it to a helper application
Added a notes section to Contributing.md

Added the s3iomain.cpp and entries to the CMakefile to compile the
GetObject helper binary.

Starter is not a Singleton anymore.

Added IOHelper::getApplicationPath() to get the absolute path of the
running application. This is used to run the helper binary.

Added the S3GetObjectProcessWrapper, which is used to run the helper
binary in a separate process and thread.

Moved the async GetObject code from S3Source to the helper binary and
made it a bit more readable.

Added unsigned (U) suffixes to some numbers in tests.